qwen3-06Btraining/requirements.txt
torch-neuronx>=2.1.0
neuronx-cc>=2.12.0
neuronx-distributed>=0.4.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0
pyyaml>=6.0
tensorboard>=2.14.0
boto3>=1.28.0
botocore>=1.31.0
requests>=2.31.0
safetensors>=0.4.0
sentencepiece>=0.1.99
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
wandb>=0.16.0